requirements.txt
Flask
mysql-connector-python
requests
pywin32
Pillow
openpyxl
pandas
